fix: resolvendo conflitos de merge
