+++ src/components/RecipeIngredients/RecipeIngredients.jsx
@@ -1,103 +1,37 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import PropTypes from 'prop-types';
+import formattingMeasuresAndIngredients from '../../services/fetchFilteredRecipes';
 
 function RecipeIngredients({ recipe, type }) {
-  const keysAndValues = Object.entries(recipe);
+  // const keysAndValues = Object.entries(recipe);
   // const [ingredients, setIngredients] = useState([]);
   // const [measures, setMeasures] = useState([]);
   const [ingredientsAndMeasures, setIngredientsAndMeasures] = useState({});
 
-  // const handleMealsIngredientsAndMeasures = useCallback(() => {
-  //   keysAndValues.map(
-  //     (element) => (
-  //       element[0].includes('strIngredient') && element[1] !== '' ? (
-  //         setMealsAndMeasures(...ingredientsAndMeasures, { ingredients: element[1] })
-  //       ) : ''),
-  //   );
-  //   keysAndValues.map(
-  //     (el) => (el[0].includes('strMeasure') && el[1] !== '' ? (
-  //       setMealsAndMeasures(...ingredientsAndMeasures, { measures: el[1] })
-  //     ) : ''),
-  //   );
-  // }, [ingredientsAndMeasures, keysAndValues]);
-
-  // const handleDrinksIngredientsAndMeasures = useCallback(() => {
-  //   keysAndValues.map(
-  //     (element) => (
-  //       element[0].includes('strIngredient') && element[1] !== null ? (
-  //         setMealsAndMeasures(...ingredientsAndMeasures, { ingredients: element[1] })
-  //       ) : ''),
-  //   );
-  //   keysAndValues.map(
-  //     (el) => (el[0].includes('strMeasure') && el[1] !== null ? (
-  //       setMealsAndMeasures([...ingredientsAndMeasures].concat({ measures: el[1] }))
-  //     ) : ''),
-  //   );
-  // }, [ingredientsAndMeasures, keysAndValues]);
-
-  // const handleIngredientsAndMeasuresSwitch = useCallback(() => {
-  //   switch (type) {
-  //   case 'meals':
-  //     handleMealsIngredientsAndMeasures();
-  //     break;
-  //   case 'drinks':
-  //     handleDrinksIngredientsAndMeasures();
-  //     break;
-  //   default:
-  //     break;
-  //   }
-  // }, [handleDrinksIngredientsAndMeasures, handleMealsIngredientsAndMeasures, type]);
-
   // filtra os valores e retorna um objeto com todos os ingredientes e medidas
-  const filterIngredientsAndMeasures = useCallback(() => {
-    const mapingValues = keysAndValues.reduce((acc, cur) => {
-      const objeto = acc;
-      const currentItem = cur;
-      if (currentItem[0].includes('strIngredient') && currentItem[1] !== '') {
-        const [itemOne, itemTwo] = currentItem;
-        objeto[itemOne] = itemTwo;
-      }
-      if (currentItem[0].includes('strMeasure') && currentItem[1] !== '') {
-        const [itemOne, itemTwo] = currentItem;
-        objeto[itemOne] = itemTwo;
-      }
-      console.log(Object.entries(objeto));
-      return objeto;
-    }, {});
-    setIngredientsAndMeasures(mapingValues);
-    return mapingValues;
-  }, [keysAndValues]);
-
-  // const comparingReduce = useCallback(() => {
-  //   console.log(ingredientsAndMeasures);
-  // }, [ingredientsAndMeasures]);
+  // const filterIngredientsAndMeasures = useCallback(() => {
+  //   const formatting = formattingMeasuresAndIngredients(type, keysAndValues);
+  //   setIngredientsAndMeasures(formatting);
+  // }, [keysAndValues, type]);
 
   useEffect(() => {
-    filterIngredientsAndMeasures();
+    const recipeskeysAndValues = Object.entries(recipe);
+    const formatting = formattingMeasuresAndIngredients(type, recipeskeysAndValues);
+    setIngredientsAndMeasures(formatting);
     // console.log(ingredientsAndMeasures);
-  }, []);
+  }, [recipe, type]);
+
+  const keyIngr = 'ingredients';
 
   return (
     <div>
       <h3>Ingredientes</h3>
       <ul>
-        { type === 'drinks'
-          ? (Object.entries(ingredientsAndMeasures).map((element, index) => (
-            (element[0].includes('strIngredient')) && element[1] !== null ? (
-              <li key={ element } data-testid={ `${index}-ingredient-name-and-measure` }>
-                { element[1] }
-              </li>
-            ) : (
-              ' '
-            ))))
-          : (Object.entries(ingredientsAndMeasures).map((element, index) => (
-            (element[0].includes('strIngredient')) && element[1] !== '' ? (
-              <li key={ element } data-testid={ `${index}-ingredient-name-and-measure` }>
-                { element[1] }
-              </li>
-            ) : (
-              ' '
-            ))))}
+        {(ingredientsAndMeasures.length > 0
+        && ingredientsAndMeasures.measures.map((element, index) => (
+          <li key={ element } data-testid={ `${index}-ingredient-name-and-measure` }>
+            { element[keyIngr] }
+          </li>)))}
       </ul>
     </div>
   );
